requirements.txt
joplin-api
commonmark
Pillow